requirements.txt
discord.py-self
httpx==0.27.0
aiofiles==23.2.0
python-dotenv==1.0.0
cryptography==42.0.0
PyNaCl==1.5.0
